src/google_ads_exclusions/requirements.txt
google-ads
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage
jsonschema
pandas
protobuf
pyarrow